Fix for loop variables
A number of for loops used loop variables that did not match the type
of the end condition.  This could lead to wraparound or
signed/unsigned mismatches.  Probably none of these are a problem in
practice, but it's fragile code.

Discussion: https://www.postgresql.org/message-id/flat/d639aede-209f-412b-927a-d38d4848b370%40eisentraut.org

+++ contrib/pg_logicalinspect/pg_logicalinspect.c
@@ -170,7 +170,7 @@ pg_get_logical_snapshot_info(PG_FUNCTION_ARGS)
 
 		arrayelems = (Datum *) palloc(ondisk.builder.committed.xcnt * sizeof(Datum));
 
-		for (int j = 0; j < ondisk.builder.committed.xcnt; j++)
+		for (size_t j = 0; j < ondisk.builder.committed.xcnt; j++)
 			arrayelems[j] = TransactionIdGetDatum(ondisk.builder.committed.xip[j]);
 
 		values[i++] = PointerGetDatum(construct_array_builtin(arrayelems,
@@ -187,7 +187,7 @@ pg_get_logical_snapshot_info(PG_FUNCTION_ARGS)
 
 		arrayelems = (Datum *) palloc(ondisk.builder.catchange.xcnt * sizeof(Datum));
 
-		for (int j = 0; j < ondisk.builder.catchange.xcnt; j++)
+		for (size_t j = 0; j < ondisk.builder.catchange.xcnt; j++)
 			arrayelems[j] = TransactionIdGetDatum(ondisk.builder.catchange.xip[j]);
 
 		values[i++] = PointerGetDatum(construct_array_builtin(arrayelems,
