+++ src/test/modules/test_radixtree/test_radixtree.c
@@ -320,7 +320,7 @@ test_random(void)
 	/* add some random values */
 	pg_prng_seed(&state, seed);
 	keys = (TestValueType *) palloc(sizeof(uint64) * num_keys);
-	for (uint64 i = 0; i < num_keys; i++)
+	for (int i = 0; i < num_keys; i++)
 	{
 		uint64		key = pg_prng_uint64(&state) & filter;
 		TestValueType val = (TestValueType) key;
@@ -333,7 +333,7 @@ test_random(void)
 
 	rt_stats(radixtree);
 
-	for (uint64 i = 0; i < num_keys; i++)
+	for (int i = 0; i < num_keys; i++)
 	{
 		TestValueType *value;
 
@@ -348,7 +348,7 @@ test_random(void)
 	qsort(keys, num_keys, sizeof(uint64), key_cmp);
 
 	/* should not find numbers in between the keys */
-	for (uint64 i = 0; i < num_keys - 1; i++)
+	for (int i = 0; i < num_keys - 1; i++)
 	{
 		TestValueType *value;
 
@@ -410,7 +410,7 @@ test_random(void)
 	pg_prng_seed(&state, seed);
 
 	/* delete in original random order */
-	for (uint64 i = 0; i < num_keys; i++)
+	for (int i = 0; i < num_keys; i++)
 	{
 		uint64		key = pg_prng_uint64(&state) & filter;
